requirements.txt
# pip install -r requirements.txt

# base ----------------------------------------

matplotlib>=3.2.2
numpy>=1.18.5
opencv-python>=4.1.2
Pillow>=7.1.2
torch>=1.9.1 
torchvision>=0.10.1
pandas>=1.1.4
streamlit>=1.10.0 



